+++ src/pos.rs
@@ -54,7 +54,7 @@ impl Pos {
         other * self.dot(other)
     }
 
-    pub fn polar(a: i32, r: f64) -> Pos {
+    pub fn polar(a: isize, r: f64) -> Pos {
         let theta = radians(a);
         Pos(r * theta.cos(), r * theta.sin())
     }
@@ -64,7 +64,7 @@ pub fn crossprod_sign(a: Pos, b: Pos, c: Pos) -> bool {
     (a.0 - c.0) * (b.1 - c.1) - (b.0 - c.0) * (a.1 - c.1) > 0.
 }
 
-pub fn radians(a: i32) -> f64 {
+pub fn radians(a: isize) -> f64 {
     (a as f64) * PI / 180.
 }
